feat(323): 메인 레이아웃 변경
