+++ src/pages/Detail.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { useQuery } from "react-query";
-import { fetchBiases, fetchEventById } from "../apis";
+import { fetchEventById, fetchPeople } from "../apis";
 import { EventType } from "../types";
 import { StyledDetail } from "../components/detail/styles/detailStyle";
 import { EventMain, EventNearHere, GoodsInfo, TwitterInfo, Location } from "../components/detail";
@@ -16,21 +16,23 @@ const Detail = () => {
 		enabled: !!id,
 	});
 
-	const { data: name } = useQuery(["bias", data], () => fetchBiases({ id: data.biasesId[0] }), {
-		enabled: !!data,
-	});
+	const { data: people } = useQuery(["bias"], () => fetchPeople());
+
+	const getBiasName = (biasId: number) => people?.filter((p) => p.id === biasId)[0].name;
 
 	useEffect(() => {
-		if (data?.place && name) {
+		if (data?.place) {
 			setMetaTags({
 				title: "오늘의 컵홀더 | 상세보기",
-				description: `${data.place}에서 열리는 ${name}의 이벤트를 오늘의 컵홀더에서 확인해보세요!`,
+				description: `${data.place}에서 열리는 ${getBiasName(
+					data?.biasesId[0]
+				)}의 이벤트를 오늘의 컵홀더에서 확인해보세요!`,
 			});
 		}
 		return () => {
 			setMetaTags({});
 		};
-	}, [data, name]);
+	}, [data]);
 
 	useEffect(() => {
 		window.scrollTo(0, 0);
