+++ src/pages/Main.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import { BiasList, EventList } from "../components/main";
+import { BiasList, EventSection } from "../components/main";
 import { StyledMain } from "../components/main/styles/mainStyle";
 import Layout from "../shared/components/layout";
 import { setMetaTags } from "../shared/utils/metaTagHandlers";
@@ -16,7 +16,7 @@ function Main() {
 		<Layout page="main">
 			<StyledMain>
 				<BiasList />
-				<EventList />
+				<EventSection />
 			</StyledMain>
 		</Layout>
 	);
